Bugfix: made navbar partial avalible to all views
Rails looks for partials that can be rendered in the accross the whole
application in the applications folder.  When the navbar partial was
in the welcome folder, it would only show up on the welcome page. Now,
since it is in the application folder, it will show up across the application.