Add recipe sends post request to collected recipes
I am doing something nonstandard. I don't want to create a new
recipe- instead I want to reorganize the existing recipe into
a user's collected recipes.  The fact that the recipe already exists
means that I'm not doing a standard crud request, so the routing
gets a little quirky.  That's why I'm passing a url paramter in a post
request string. This seems like a stinky way to do it.