BUGFIX:  user recipes page now displays
When we extracted the recipes display logic into an application
partial, it broke the code in users_recipes controller.
Specifically, the User recipes controller was senting the partial
the instance variable @user_recipes, and the partial was expecting
an instance variable named @recipes, so all of my recipes broke.

It now displayes correctly.